requirements.txt
llama-index
streamlit
llama-index-embeddings-gemini
llama-index-llms-gemini
python-dotenv
nltk
pyaudio
pyttsx3
numpy
openai-whisper
gTTs